+++ trade/src/main/java/bisq/trade/mu_sig/messages/network/MuSigSetupTradeMessage_A.java
@@ -20,7 +20,7 @@ package bisq.trade.mu_sig.messages.network;
 import bisq.contract.ContractSignatureData;
 import bisq.contract.mu_sig.MuSigContract;
 import bisq.network.identity.NetworkId;
-import bisq.trade.mu_sig.messages.network.vo.PubKeyShares;
+import bisq.trade.mu_sig.messages.network.mu_sig_data.PubKeyShares;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
